Issue#5 "Reconnecting message visible" is resolved
Resolved "Reconnecting message should be visible in the middle of the
screen" issue.

Added a VNCEvent class and its child VNCReconnectEvent class. Moved to
VNCErrorEvent class to the aforementioned class hierarchy; it now
extends VNCEvent class.

Added ReconnectingCountDown.mxml which handles displaying of the
reconnect message as described in the issue.