Always send ctrl
This patch makes it possible to send Ctrl keys to the server, which
makes it possible to do Ctrl-click to, for example, do multiselect on
the web.

We do this by always sending ctrl. This then breaks pasting text
(Ctrl-v), becaue the first character of the pasted text is assumed by
the server to be part of the ctrl sequence. We "fix" this, by sending a
ctrl key-up just before every character in the pasted sequence.